Moved tests and fixed typs
